requirements.txt
Django>=4.0.0
djangorestframework
pyyaml
requests
django-cors-headers
psycopg2
django-filter
meteomatics
django-phonenumber-field[phonenumbers]